moves admin component tests to their own dir
