+++ backend/server/public/__tests__/admin/components_admin_addetail.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import { AdDetailUnconnected } from "../src/components/admin/addetail.jsx";
+import { AdDetailUnconnected } from "../../src/components/admin/addetail.jsx";
 
 Enzyme.configure({ adapter: new Adapter() });
