+++ backend/server/public/__tests__/admin/components_admin_ads.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import { AdsUnconnected } from "../src/components/admin/ads.jsx";
+import { AdsUnconnected } from "../../src/components/admin/ads.jsx";
 Enzyme.configure({ adapter: new Adapter() });
 
 function setup({ ads, pagination }) {
