+++ backend/server/public/__tests__/admin/components_admin_login.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import { LoginUnconnected } from "../src/components/admin/login.jsx";
+import { LoginUnconnected } from "../../src/components/admin/login.jsx";
 
 Enzyme.configure({ adapter: new Adapter() });
